requirements.txt
click==8.0.1
colorama==0.4.4
cycler==0.10.0
flask==2.0.1
flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mysqlclient==2.0.3
numpy==1.21.0
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
pandas==1.3.0
Pillow==8.3.1
prettytable==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
wcwidth==0.2.5
Werkzeug==2.0.1
